requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openpyxl==3.1.2
pandas==2.1.4
weasyprint==61.2
pydyf==0.10.0
Pillow==10.2.0
python-dotenv==1.0.0
pytz==2024.1
babel==2.14.0
jinja2==3.1.3

